Add support for input.entity.property = null comparison.
For the NullValue, Scalar#getValue() returns null.
When that is passed into Expressions.constant(value), at some point value.getClass() is called,
which produces a NullPointerException in the case that value is null.

Instead, use the special construct from QueryDSL to represent a IS NULL expression.
This construct correctly handles the translation from 'property = ?' to 'property IS NULL', which is required
for NULL-checks in the database.