Introduce `ThunkExpression#simplify()`
To create a correct representation of `ThunkExpression#canBeResolved()` and `ThunkExpression#resolve()`,
a lot of work would have to be duplicated between the two methods.

Having a strictly binary "can be resolved"/"can't be resolved" prevents
expression simplification when there is any unknown variable.

Moving from complete resolve to a method to simplify the expression as
much as possible allows partial simplification for the expression, or in
certain cases complete elimination for the variable part.

e.g.: `true OR input.entity.color == "blue"` can now be simplified to `true`

Simplification also allows the removal of unnecessary terms.

e.g.: `true AND input.entity.color == "blue"` can now be simplified to `input.entity.color == "blue"`