Replace path-based domain type lookup with using the parameter
`RootResourceInformationHandlerMethodArgumentResolver#postProcess()`
gets the domain type passed as a parameter. Instead of trying to extract
the entity class from the ResourceMappings ourselves, we can just use
this parameter value.

As a bonus, the abac context is now always set when there is an
X-ABAC-Context header, allowing non-generated API endpoints to use the
ABAC context as well.