Ensure same variable name is used in WHERE clause as in SELECT/FROM
Most of the time, QueryDSL `PathBuilderFactory` uses the same variable
name as in the generated `Q*` classes (which is used in select/from by
spring-data-rest using the `EntityPathResolver`). However, this breaks
down when an entity is used that would have a reserved keyword as simple
name.

In that case, the generated `Q*` class has a '1' appended to the
variable (to avoid using a reserved keyword), but the
`PathBuilderFactory` does not. (See
https://github.com/querydsl/querydsl/issues/1891 for the rationale).

We must use the same strategy for naming our predicates as used by
spring-data-rest to make this work, so a Thunx `PathBuilderFactory`
interface is introduced, together with an implementation that creates a
`PathBuilder` based on the metadata from the spring-data-rest `EntityPathResolver`.