Introduce QuerydslPredicateResolver interface & beans
Because it is difficult to extend the
RootResourceInformationHandlerMethodArgumentResolver further with
additional customizations with additional QueryDSL Predicates while
preserving existing predicates.
Registering a `QuerydslPredicateResolver` bean allows adding an
additional predicate into the mix, it automatically being AND'ed
together with other supplied predicates.
It is also possible to suppress or modify some predicates by using a
`BeanPostProcessor` to modify or wrap existing beans.