Json Encoder implements ThunkExpressionEncoder & ThunkExpressionDecoder interfaces
Instead of writing ad-hoc functions for implementing ThunkExpressionEncoder and ThunkExpressionDecoder using the ExpressionJsonConverter,
replace it with a new JsonThunkExpressionCoder that directly implements both interfaces.

To maintain backwards compatibility, it is not possible for ExpressionJsonConverter to implement those interfaces, as methods with the
same name and parameter types can not have a different return type.
Because of this, ExpressionJsonConverter is now deprecated for removal in the next major version.