Fix 'object references an unsaved transient instance' error [ACC-1827]
By adding FetchType.LAZY to @OneToOne and @ManyToOne relations, their references are no longer automatically updated when CrudRepository.save() returns a different instance (usually when it is not yet present in the db). When fetching such an entity with an old reference, Hibernate will not find the id of the target entity and throws 'object references an unsaved transient instance' error. The problem is solved by setting the @OneToOne or @ManyToOne relation manually after the other side was saved.